tags, category models moved to their correct folders
