+++ index.js
@@ -1,7 +1,6 @@
 const server = require('./api/server');
 
 const port = process.env.PORT || 6000;
-
 server.listen(port, () => {
   console.log('listening on port ' + port);
 });
