+++ rest-api/category/routes/category-routes.js
@@ -1,7 +1,8 @@
 const router = require('express').Router();
 
-const { getCategories } = require('../controllers/category-controllers');
+const { getCategories, editCategory } = require('../controllers/category-controllers');
 
 router.get('/', getCategories);
+router.put('/:id/categories', editCategory);
 
 module.exports = router;
