+++ rest-api/recipe/controllers/recipe-controllers.js
@@ -5,9 +5,7 @@ module.exports = {
   getRecipes,
   getRecipeById,
   addRecipe,
-  editRecipeInfo,
-  editTag,
-  editCategory
+  editRecipeInfo
 };
 
 async function getRecipes(req, res) {
@@ -67,34 +65,4 @@ async function editRecipeInfo(req, res) {
   }
 }
 
-async function editTag(req, res) { 
-  const { id } = req.params; 
-  try { 
-    const editTag = await dbRecipe.editTag( id, req.body);
-    editTag 
-    ? res.status(200).json({ message: 'tags succesffully updated'})
-    : res.status(401).json({ message: 'tag id does nont match '})
-  }
-  catch(error) { 
-    res.status(500).json({ 
-      message: `update unsuccesful for id` + id, 
-      error
-    })
-  }
-}
 
-async function editCategory(req, res){ 
-  const { id } = req.params;
-  try{ 
-    const editCategory= await dbRecipe.editCategory( id,req.body);
-    editCategory 
-    ? res.status(200).json({ message: 'tags succesffully updated'})
-    : res.status(401).json({ message: 'tag id does not match '})
-  }
-  catch(error) { 
-    res.status(500).json({ 
-      message: `update unsuccesful for id` + id, 
-      error
-    })
-  }
-}
