+++ rest-api/recipe/models/recipe-models.js
@@ -5,35 +5,15 @@ module.exports = {
   getRecipeById,
   addRecipeTransaction, 
   editRecipeInfo,
-  editTag,
   editCategory
 };
 
 //sam
-// async function editRecipeInfo (id , recipeUpdate) { 
-//   const editRecipeInfo = await db('recipes')
-//   .join('users','users.id', 'recipes.user_id')
-//   .select(
-//     'users.id',
-//     'recipes_id',
-//     'recipes.parent_id',
-//     'recipes.title as title',
-//     'recipes.description',
-//     'recipes.time_required',
-//     'recipes.difficulty',
-//     'recipes.budget'
-//   )
-//   .where('recipes.id', id )
-//   .update(recipeUpdate)
-//   return editRecipeInfo
-// }
 
 async function editRecipeInfo(id, body) { 
   const { title, description, time_required, difficulty, budget, parent_id } = body;
   return  await db.transaction(async trx => { 
     try { 
-    
-        
         await trx('recipes')
         .update('title', title)
         .where('recipes.id', id)
@@ -68,80 +48,6 @@ async function editRecipeInfo(id, body) {
     }
   });
 };
-
-
-async function editTag(recipe_id, body) { 
-
-  const { name } = body;
-
-  return await db.transaction( async trx => { 
-    try {
-      const [tags_name] = await trx('tags')
-      .where('tags.name', name)
-      .select('tags.id');
-
-      const tags_id = tags_name ? tags_name.id : undefined;
-      const recipeTagsId = await trx('recipe_tags')
-      .where('recipe_tags.recipe_id', recipe_id)
-      .select('recipe_tags.id')
-
-      const { id } = recipeTagsId[index] ? recipeTagsId[index] : { id : undefined }
-      if (!id) { throw 'index is out of range!'};
-
-      if(tags_id) { 
-        await trx('recipe_tags')
-        .update('tags_id', tags_id)
-        .where('id', id )
-      } else{ 
-        const [ newTagId] = await trx('tags')
-        .insert({ name })
-        .returning('id');
-
-        await trx('recipe_tags')
-          .update('tags_id', newTagId)
-          .where('id', id)
-      }
-      return await trx('recipe_tags')
-      .where('recipe_tags.recipe_id', recipe_id)
-      .select('tags.name');
-    } catch(error) { 
-      console.log(error);
-      throw(error)
-    }
-  })
-}
-
-
-async function editCategory(recipe_id, body) { 
-  const { name } = body; 
-  return await db.transaction( async trx => { 
-    try{ 
-      const [categories_name] = await trx('categories')
-      .where('categories.name', name)
-      .select('categories.id');
-
-      const categories_id = categories_name ? categories_name.id : undefined;
-      const recipeCategoriesId = await trx('recipes_categories')
-      .where('recipes_categories.recipe_id', recipe_id)
-      .select('recipe_categories.id')
-
-      const { id } = recipeCategoriesId[index] ? 
-      recipeCategoriesId[index] : { id : undefined }
-      if (!id ) { throw 'index is out of range!'};
-    
-      await trx( 'recipe_categories')
-      .update('categories_id', categories_id)
-      .where( 'id', id)
-
-      return await trx('recipe_categories') 
-      .where('recipe_categories.recipe_id', recipe_id)
-      .select('categories.name')
-    } catch(error) { 
-      console.log(error)
-      throw(error)
-    }
-  })
-}
 
 async function getRecipes() {
   const recipes = await db('recipes')
