+++ rest-api/tag/routes/tag-routes.js
@@ -1,7 +1,11 @@
 const router = require('express').Router();
 
-const { getTags } = require('../controllers/tag-controllers');
+const { getTags,
+    editTag
+} = require('../controllers/tag-controllers');
 
 router.get('/', getTags);
+router.put('/:id/tags', editTag)
+
 
 module.exports = router;
